Fix: turn off invariant globalization to fix sql database invariant culture exception
The Invariant Globalization setting must be explicitly set to `false` (not just omitted) when running in a chiseled or alpine container, because the containers lack certain culture and locale information. The exception message ("Only the invariant culture is supported in globalization-invariant mode") and the Microsoft link below imply that Invariant Globalization must be turned on, but in fact it's SQL Database throwing an exception because it expects *more* cultures than just the invariant culture to be available.

https://github.com/dotnet/SqlClient/issues/2239

https://aka.ms/GlobalizationInvariantMode