Replace dustytables with Npgsql.FSharp
Npgsql.FSharp has an API that's nearly identical to dustytables (I'm
guessing that, in reality, dustytables actually has an API that's
identical to Npgsql.FSharp). Since Dustytables doesn't actually work
with Postgres, this move was both a no-brainer and a fix to get the
database modules working with the postgresql database.

type: fix
scope: dependencies, database