Add Migrator `down` verb
FluentMigrations doesn't (easily) support a `to X` verb where `X` is a
migration that's below the current migration, i.e. it does not easily
support going down in a targeted manner because it doesn't know which
migration the database is on before it runs. We work around this
limitation by letting the CLI specify which direction it should go
explicitly.

type: feature
scope: migrations, database